move controls back into their original location
