+++ src/plugins/dataflow/nodes/controls/plot-button-control.scss
