+++ src/plugins/dataflow/rete/nodes/sensor-node.ts
@@ -3,8 +3,8 @@ import { typeField } from "../../../../utilities/mst-utils";
 import { BaseNode, BaseNodeModel, NoInputs } from "./base-node";
 import { INodeServices } from "../service-types";
 import { ClassicPreset } from "rete";
-import { DropdownListControl, IDropdownListControl, ListOption } from "../controls/dropdown-list-control";
-import { PlotButtonControl } from "../controls/plot-button-control";
+import { DropdownListControl, IDropdownListControl, ListOption } from "../../nodes/controls/dropdown-list-control";
+import { PlotButtonControl } from "../../nodes/controls/plot-button-control";
 import { numSocket } from "../num-socket";
 import { NodeSensorTypes, kSensorMissingMessage, kSensorSelectMessage } from "../../model/utilities/node";
 import { NodeChannelInfo, kDeviceDisplayNames } from "../../model/utilities/channel";
